feat: implement all tensor_scalar ALU ops and types
This patch fills in the implementation of tensor_scalar, and also
refactors some code to make this easier. The main change is the
introduction of specific element types (`Dtype`) as specified in the
NKI docs.

+++ interop/test/test_basic.py
@@ -142,7 +142,7 @@ def undefined_ok(t):
   undefined_ok
   ])
 def test_succeed(f):
-  t = np.ndarray(10)
+  t = np.ndarray(10, dtype="float32")
   F = Parser(f)   # parse python
   F(t)            # specialize, and reduce to KLR
